feat: In case a blacklisted user wins an event, the bot won't give the user anything.

+++ src/utils/Microservice.ts
@@ -4,29 +4,38 @@ import path from "path";
 export namespace Microservice {
   export async function ListenForWinner(): Promise<void> {
     setInterval(async () => {
-        const winnerPath: string = path.join(process.cwd(), "winners.json")
+      const winnerPath: string = path.join(process.cwd(), "winners.json");
       if (fs.existsSync(winnerPath)) {
-        console.log("Winner file exists")
+        console.log("Winner file exists");
         delete require.cache[require.resolve(winnerPath)];
         const winner = require(winnerPath);
         const winnerId = String(winner.id);
         const winnerAmount = winner.amount;
         if (await GoodieController.findUser(winnerId)) {
-          await GoodieController.incrementCandy(winnerId, winnerAmount).then(() => {
-            fs.unlink(winnerPath, (e) => {
+          if (await GoodieController.isBlackListed(winnerId)) {
+            return fs.unlink(winnerPath, (e) => {
+              console.log("Winners file deleted");
+            });
+          }
+          await GoodieController.incrementCandy(winnerId, winnerAmount).then(
+            () => {
+              fs.unlink(winnerPath, (e) => {
                 console.log("Winners file deleted");
               });
-          })
+            }
+          );
         } else {
           await GoodieController.createUser(winnerId);
-          await GoodieController.incrementCandy(winnerId, winnerAmount).then(() => {
-            fs.unlink(winnerPath, (e) => {
+          await GoodieController.incrementCandy(winnerId, winnerAmount).then(
+            () => {
+              fs.unlink(winnerPath, (e) => {
                 console.log("Winners file deleted");
               });
-          })
+            }
+          );
         }
       } else {
-        console.log("Not found")
+        console.log("Not found");
         return;
       }
     }, 5000);
